move BrokenSiteSender interface to broken-site-api module
